requirements.txt
Django==1.6.10
South==1.0.2
PyYAML==3.11
